+++ lib/locatine/find_logic.rb
@@ -11,11 +11,19 @@ module Locatine
       name
     end
 
+    def stable?(attributes)
+      s = []
+      attributes.each_pair do |_depth, array|
+        s.push (array.max_by {|item| item['stability'].to_i})['stability'].to_i
+      end
+      s.max > 1
+    end
+
     def core_search(result, name, scope, vars, exact)
       if @data[scope][name].to_h != {}
         result = find_by_data(@data[scope][name], vars)
         attributes = generate_data(result, vars) if result
-        if !result && !exact
+        if !result && (!exact || !stable?(@data[scope][name]))
           result, attributes = find_by_magic(name, scope,
                                              @data[scope][name], vars)
         end